Use long long for get_malloc_size_host() get_malloc_size_dev()

+++ src/cublas/cublas_create_allocate.cpp
@@ -63,10 +63,10 @@ long get_malloc_size_scalar(mblas_data_type type) {
   return type_call_host<sizeofCUDT>(type);
 }
 
-long get_malloc_size_host(mblas_data_type type, long x, long y, int batch, long long stride) {
+long long get_malloc_size_host(mblas_data_type type, long x, long y, int batch, long long stride) {
   int typesize = type_call_host<sizeofCUDT>(type);
-  long base = x * y;
-  long total_elements;
+  long long base = x * y;
+  long long total_elements;
   if (batch > 1 && stride > base) {
     total_elements = stride * (batch - 1) + base;
   } else {
@@ -75,11 +75,11 @@ long get_malloc_size_host(mblas_data_type type, long x, long y, int batch, long 
   return total_elements * typesize;
 }
 
-long get_malloc_size_dev(mblas_data_type type, long x, long y, int batch, long long stride) {
+long long get_malloc_size_dev(mblas_data_type type, long x, long y, int batch, long long stride) {
   int typesize = type_call_dev<sizeofCUDT>(type);
   long packing_count = type.get_packing_count();
-  long base = x * y;
-  long total_elements;
+  long long base = x * y;
+  long long total_elements;
   if (batch > 1 && stride > base) {
     total_elements = stride * (batch - 1) + base;
   } else {
